requirements.txt
django>=2.0
pysimplemodel==2.3.3
requests
unipath
